# REST API/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "idk",
        "description": "Assignment #1: gRPC and REST API Implementation by idk",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "paths": {
...
        "/book/insert": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Insert new book",
                "parameters" : [
                    {
                        "name": "book",
                        "in": "body",
                        "description": "Book that we want to insert",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New book is inserted",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Book": {
            "required": [
                "id",
                "title",
                "author"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                }
            }
        }
    },
    "schemes": [
            "http"
        ],
        "consumes": [
            "application/json"
        ],
        "produces": [
            "application/json"
    ]
}